requirements.txt
aiofiles==23.2.1
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.16.0
click==8.1.7
cryptography==39.0.2
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.103.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.4
passlib==1.7.4
pkg_resources==0.0.0
psycopg2-binary==2.9.6
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyMySQL==1.0.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.20.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
zipp==3.17.0
